requirements.txt
gradio==5.4.0
numpy==2.1.2
pydantic==2.9.2
python-dotenv==1.0.1
python_ffmpeg==2.0.12
torch==2.6.0
torchaudio==2.6.0
transformers==4.49.0
accelerate==1.5.1
bitsandbytes==0.44.1
FlagEmbedding==1.3.4
chromadb==0.6.2
ollama==0.4.7
ffmpeg==1.4
elevenlabs==1.54.0
openai==1.68.2
mxbai-rerank==0.1.3
pdfplumber==0.11.6